classes: introduce new bbclass to handle package version generation
Introduce new class that provides a functionality to record the version
of the component in the format, that could be used during sw-description
file construction.

Once this class is inherited by a recipe - it would generate a file in
image deployment folder, which would contain a version of the package
(either SRCREV or PV) together with a package name.

Signed-off-by: Andrey Zhizhikin <andrey.zhizhikin@leica-geosystems.com>